+++ packages/bytes/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "bytes/5.0.0-beta.128";
+export declare const version = "bytes/5.0.0-beta.130";