+++ packages/keccak256/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "keccak256/5.0.0-beta.126";
+export declare const version = "keccak256/5.0.0-beta.128";