+++ packages/pbkdf2/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "pbkdf2/5.0.0-beta.126";
+export declare const version = "pbkdf2/5.0.0-beta.128";