+++ packages/solidity/lib/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "solidity/5.0.0-beta.125";
+export declare const version = "solidity/5.0.0-beta.127";