+++ packages/strings/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "strings/5.0.0-beta.129";
+export declare const version = "strings/5.0.0-beta.131";