+++ packages/transactions/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "transactions/5.0.0-beta.127";
+export declare const version = "transactions/5.0.0-beta.129";